+++ engine/tests/new-tests/sample-generator.cu
@@ -1,7 +1,7 @@
-#include <from_cudf/cpp_tests/utilities/base_fixture.hpp>
-#include <from_cudf/cpp_tests/utilities/column_utilities.hpp>
-#include <from_cudf/cpp_tests/utilities/column_wrapper.hpp>
-#include <from_cudf/cpp_tests/utilities/type_lists.hpp>
+#include <tests/utilities/base_fixture.hpp>
+#include <tests/utilities/column_utilities.hpp>
+#include <tests/utilities/column_wrapper.hpp>
+#include <tests/utilities/type_lists.hpp>
 
 #include <execution_graph/logic_controllers/LogicPrimitives.h>
 
@@ -13,7 +13,7 @@ class SampleGeneratorTest : public BlazingUnitTest {};
 
 using NumericTypesForSampling = cudf::test::Types<int8_t, int16_t, int32_t, int64_t, float, double>;
 
-TYPED_TEST_SUITE(SampleGeneratorTest, NumericTypesForSampling);
+TYPED_TEST_CASE(SampleGeneratorTest, NumericTypesForSampling);
 
 TYPED_TEST(SampleGeneratorTest, BaseCase) {
 	using T =  TypeParam;